requirements.txt
py-cord>=2.6.0
aiohttp>=3.9.0
openai>=1.0.0
python-dotenv>=1.0.0
yfinance
feedparser
